Update max and default values type

+++ src/main/resources/sql/postgresql/Migration_4.0.0_to_4.1.0.sql
@@ -97,8 +97,8 @@ UPDATE policy SET system = false, status = true, default_status = true WHERE id 
 
 -- Add new field to Functionality_integer and Functionality_unit
 
-ALTER TABLE functionality_unit ADD integer_default_value int8 NULL;
-ALTER TABLE functionality_integer ADD integer_default_value int8 NULL;
+ALTER TABLE functionality_unit ADD integer_default_value int4 NULL;
+ALTER TABLE functionality_integer ADD integer_default_value int4 NULL;
 
 ALTER TABLE functionality_unit RENAME COLUMN integer_value TO integer_max_value;
 ALTER TABLE functionality_integer RENAME COLUMN integer_value TO integer_max_value;
